Plusieurs avancements

+++ assignments/src/app/assignments/assignment-detail/assignment-detail.component.ts
@@ -1,19 +1,24 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, inject, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { AssignmentsService } from 'src/app/shared/assignments.service';
 import { Assignment } from '../assignment.model';
+import {MatSnackBar, MatSnackBarRef} from '@angular/material/snack-bar';
 
 @Component({
   selector: 'app-assignment-detail',
   templateUrl: './assignment-detail.component.html',
   styleUrls: ['./assignment-detail.component.css'],
 })
-export class AssignmentDetailComponent implements OnInit {
-  assignmentTransmis!: Assignment|undefined ;
 
-  constructor(private assignmentsService: AssignmentsService,
-              private route:ActivatedRoute,
-              private router:Router) {}
+export class AssignmentDetailComponent implements OnInit {
+  assignmentTransmis!: Assignment | undefined;
+
+  constructor(
+    private assignmentsService: AssignmentsService,
+    private route: ActivatedRoute,
+    private router: Router,
+    private _snackBar: MatSnackBar
+  ) {}
 
   // Appelé AVANT l'affichage du composant, fait partie du
   // cycle de vie du composant
@@ -21,12 +26,17 @@ export class AssignmentDetailComponent implements OnInit {
     this.getAssignment();
   }
 
+  openSnackBar(message:string) {
+    this._snackBar.open(message, "Fermer", {
+      duration: 3000
+    });
+  }
+
   getAssignment() {
     // on récupère l'id dans l'url
     // Le + force la conversion en number
-    const id:number = +this.route.snapshot.params['id'];
-    this.assignmentsService.getAssignment(id)
-    .subscribe((assignment) => {
+    const id: number = +this.route.snapshot.params['id'];
+    this.assignmentsService.getAssignment(id).subscribe((assignment) => {
       this.assignmentTransmis = assignment;
     });
   }
@@ -38,22 +48,32 @@ export class AssignmentDetailComponent implements OnInit {
       .updateAssignment(this.assignmentTransmis)
       .subscribe((message) => {
         console.log(message);
-        // et on navigue vers la page d'accueil qui affiche
-        // la liste des assignments
-        this.router.navigate(["/home"]);
+        this.router.navigate(['/home']);
       });
   }
 
+  modifierAssignment(devoir: Assignment) {
+    this.router.navigate(['/assignment/' + devoir.id + '/edit']);
+  }
+
   onDeleteAssignment() {
-    if (!this.assignmentTransmis) return;
-    this.assignmentsService
-      .deleteAssignment(this.assignmentTransmis)
-      .subscribe((reponse) => {
-        console.log(reponse.message);
-        this.assignmentTransmis = undefined;
-        // et on navigue vers la page d'accueil qui affiche
-        // la liste des assignments
-        this.router.navigate(["/home"]);
-      });
+    var result = confirm('Etes-vous sûr de vouloir supprimer ce devoir ?');
+
+    if (result == true) {
+      if (!this.assignmentTransmis) return;
+
+      this.assignmentsService
+        .deleteAssignment(this.assignmentTransmis)
+        .subscribe((reponse) => {
+          console.log(reponse.message);
+          this.assignmentTransmis = undefined;
+          // et on navigue vers la page d'accueil qui affiche
+          // la liste des assignments
+          this.router.navigate(['/home']);
+        });
+    } else {
+      this.openSnackBar('Suppression annulée');
+      return;
+    }
   }
 }
